Fix mushroom log and mossy seed effect ranges
Mushroom log and mossy seed ranges added in b4d3a876 wrongly
display quadruple (resp. double the square radius) of their actual size.

Wiki pages
https://stardewvalleywiki.com/Mushroom_Log
and
https://stardewvalleywiki.com/Moss#Moss_Growth
describe the areas as 7x7 and 5x5 respectively, so the
`maxDisplaySquareRadius` value needs to be half of the area size
minus one. This is also consistent with my own observation of the
mushroom log behavior.